requirements.txt
Django==5.0
django-crispy-forms==2.3
Pillow==10.1.0
python-dotenv==1.0.0
django-environ==0.11.2
django-widget-tweaks==1.5.0
django-debug-toolbar==4.2.0
psycopg2-binary==2.9.9
django-storages==1.14.2
gunicorn==21.2.0
